refactor(structure): Tidy settings' components
