+++ src/components/Form/Settings/Slider/Slider.jsx
