+++ src/components/Layout/Settings/Section/Section.jsx
@@ -1,7 +1,7 @@
 import { MdOutlineKeyboardArrowRight } from 'react-icons/md';
 import React from 'react';
 
-export default function Section({ title, subtitle, icon, onClick, children }) {
+function Section({ title, subtitle, icon, onClick, children }) {
   return (
     <div className="moreSettings" onClick={onClick}>
       <div className="left">
@@ -12,8 +12,10 @@ export default function Section({ title, subtitle, icon, onClick, children }) {
         </div>
       </div>
       <div className="action">
-      {React.Children.count(children) === 0 ? <MdOutlineKeyboardArrowRight /> : children}
+        {React.Children.count(children) === 0 ? <MdOutlineKeyboardArrowRight /> : children}
       </div>
     </div>
   );
 }
+
+export { Section as default, Section };
