+++ src/components/Layout/Settings/Section/index.jsx
@@ -1,0 +1,1 @@
+export * from './Section';
