+++ src/components/Layout/WidgetsLayout/WidgetsLayout.jsx
@@ -1,0 +1,5 @@
+const WidgetsLayout = ({ children }) => {
+  return <div id="widgets">{children}</div>;
+};
+
+export { WidgetsLayout as default, WidgetsLayout };
