+++ src/components/Layout/WidgetsLayout/index.jsx
@@ -1,0 +1,1 @@
+export * from './WidgetsLayout';
