+++ src/components/Layout/index.jsx
@@ -1,0 +1,1 @@
+export * from './WidgetsLayout';
