+++ src/features/modals/main/marketplace/Item.jsx
@@ -17,7 +17,7 @@ import {
 } from 'react-icons/md';
 import Modal from 'react-modal';
 
-import Header from '../settings/Header';
+import Header from '../../../../components/Layout/Settings/Header/Header';
 import { Button } from 'components/Elements';
 
 import { install, uninstall } from 'modules/helpers/marketplace';
