+++ src/features/modals/main/settings/sections/Date.jsx
@@ -1,11 +1,11 @@
 import variables from 'config/variables';
 import { useState } from 'react';
 
-import Header from '../Header';
-import Checkbox from '../Checkbox';
-import Dropdown from '../Dropdown';
+import Header from '../../../../../components/Layout/Settings/Header/Header';
+import Checkbox from '../../../../../components/Form/Settings/Checkbox/Checkbox';
+import Dropdown from '../../../../../components/Form/Settings/Dropdown/Dropdown';
 
-import { Row, Content, Action } from '../SettingsItem';
+import { Row, Content, Action } from '../../../../../components/Layout/Settings/Item/SettingsItem';
 import PreferencesWrapper from '../PreferencesWrapper';
 
 export default function Date() {
